managed to read a macho header using the memory manipulation functions, good sign :)

+++ memory/src/memory_map/cast_memory.c
@@ -15,12 +15,12 @@
 static t_memory_error	conv_integer_memory(uint8_t *mem_out,
 	uint8_t *mem_in, t_memory_descriptor cmd[2], uint32_t endian_mode)
 {
-	uint8_t	pad;
-	size_t	sw;
+	uint8_t		pad;
+	uint64_t	sw;
 
 	pad = (cmd[0].sign && (mem_in[!(endian_mode & 1)
 		? cmd[0].block_size - 1 : 0] & 0x80)) * 0xFF;
-	sw = (size_t)-1;
+	sw = (uint64_t)-1;
 	if (2 == endian_mode)
 		while (++sw < cmd[1].block_size)
 			mem_out[sw] = cmd[1].block_size - sw - 1 < cmd[0].block_size
@@ -43,19 +43,19 @@ static t_memory_error	conv_integer_memory(uint8_t *mem_out,
 t_memory_error			cast_memory(t_memory_map *out, t_memory_map *in,
 	t_memory_descriptor md_out, t_memory_descriptor md_in)
 {
-	size_t	align_in;
-	size_t	align_out;
-	size_t	it;
-	size_t	j;
+	uint64_t	align_in;
+	uint64_t	align_out;
+	uint64_t	it;
+	uint64_t	j;
 
 	if (valid_cursor(out, md_out, &align_out) != ME_OK)
 		return (ME_PENDING_ERROR);
 	if (valid_cursor(in, md_in, &align_in) != ME_OK)
 		return (ME_PENDING_ERROR);
-	it = (size_t)-1;
+	it = (uint64_t)-1;
 	while (++it < md_out.nb_blocks)
 	{
-		j = (size_t)-1;
+		j = (uint64_t)-1;
 		while (++j < align_out)
 			out->ptr[it * align_out + j + out->cursor] = 0;
 		if (it < md_in.nb_blocks)
